add common utils
